cast both ints to doubles for the X and Y scale factors
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@57400 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/common/dcsvg.cpp
@@ -579,7 +579,7 @@ void wxSVGFileDCImpl::NewGraphics ()
     }
 
     sLast.Printf( wxT("stroke-width:%d\" \n   transform=\"translate(%.2g %.2g) scale(%.2g %.2g)\">"),
-                  w, (double)m_logicalOriginX, m_logicalOriginY, m_scaleX, m_scaleY  );
+                  w, double(m_logicalOriginX), double(m_logicalOriginY), m_scaleX, m_scaleY  );
 
     s = sBrush + sPenCap + sPenJoin + sPenStyle + sLast + newline + sWarn;
     write(s);
